+++ CHANGELOG.md
@@ -7,9 +7,15 @@ Focus was on adding more functionality, and modernizing the package setup.
 - Added tests for ```paths```, ```type_conversions``` and ```validation```
 - Removed ```type_conversion.stringify_list``` since functionality can be one-lined using ```str.join```
 - Added ```type_conversion.dict_to_defaultdict```; Takes in a dictionary, and a default Callable then converts the dict to a defaultdict
+- Changed From Apache V2 to MIT liscence
+- Added ```paths.add_to_path``` for adding specified paths to system paths
+- Added testing for ```sdu.paths```, ```sdu.validation```, ```sdu.autocomplete```, and ```sdu.type_conversions```
+- Changed ```sdu.terminal``` to ```sdu.cli```
 
 ### Documentation Improvements
 - Changed to second point releases (from 0.0.x to 0.x.0)
+- Improved the index documentation to be more clear and concise of an overview
+- Added bullet points to each package for a quick overview of what each module can be used for
 
 ### Bug Fixes
 - Fixed some preprocessing bugs with windows in ```path.preprocess_paths()```